+++ neurocaps/_utils/_check_parcel_approach.py
@@ -1,5 +1,5 @@
 """Internal function for checking the validity of parcel_approach."""
-import copy, os, re, warnings
+import copy, os, re, textwrap, warnings
 from nilearn import datasets
 
 def _check_parcel_approach(parcel_approach, call = "TimeseriesExtractor"):
@@ -15,22 +15,22 @@ def _check_parcel_approach(parcel_approach, call = "TimeseriesExtractor"):
                                                                  "rh": [5]}}}}
 
     if not isinstance(parcel_approach,dict) or isinstance(parcel_approach,dict) and len(parcel_approach) > 0 and not isinstance(parcel_approach[list(parcel_approach)[0]],dict):
-        raise ValueError(f"""
+        raise ValueError(textwrap.dedent(f"""
                          Please include a valid `parcel_approach` in one of the following dictionary
                          formats for 'Schaefer' or 'AAL' {valid_parcel_dict}
-                         """)
+                         """))
 
     if len(parcel_approach) > 1:
-        raise ValueError(f"""
+        raise ValueError(textwrap.dedent(f"""
                          Only one parcellation approach can be selected.
                          Example format of `parcel_approach`: {valid_parcel_dict}
-                         """)
+                         """))
 
     if "Schaefer" not in parcel_approach and "AAL" not in parcel_approach and "Custom" not in parcel_approach:
-        raise ValueError(f"""
+        raise ValueError(textwrap.dedent(f"""
                          Please include a valid `parcel_approach` in one of the following formats for
                          'Schaefer', 'AAL', or 'Custom': {valid_parcel_dict}
-                         """)
+                         """))
 
     if "Schaefer" in parcel_approach:
         if "n_rois" not in parcel_approach["Schaefer"]:
@@ -72,30 +72,30 @@ def _check_parcel_approach(parcel_approach, call = "TimeseriesExtractor"):
 
     if "Custom" in parcel_approach:
         if call  == "TimeseriesExtractor" and "maps" not in parcel_approach["Custom"]:
-            raise ValueError(f"""
+            raise ValueError(textwrap.dedent(f"""
                              For `Custom` parcel_approach, a nested key-value pair containing the key 'maps' with the
                              value being a string specifying the location of the parcellation is needed.
                              Example: {valid_parcel_dict['Custom']}
-                             """)
+                             """))
         check_subkeys = ["nodes" in parcel_approach["Custom"], "regions" in parcel_approach["Custom"]]
         if not all(check_subkeys):
             missing_subkeys = [["nodes", "regions"][x] for x,y in enumerate(check_subkeys) if y is False]
             error_message = f"The following sub-keys haven't been detected {missing_subkeys}"
             if call == "TimeseriesExtractor":
-                warnings.warn(f"""
+                warnings.warn(textwrap.dedent(f"""
                               {error_message}.
                               These labels are not needed for timeseries extraction but are needed for future
-                              timeseries or CAPs plotting.""")
+                              timeseries or CAPs plotting."""))
             else:
                 custom_example = {"Custom": {"nodes": ["LH_Vis1", "LH_Vis2", "LH_Hippocampus",
                                                        "RH_Vis1", "RH_Vis2", "RH_Hippocampus"],
                                              "regions": {"Vis" : {"lh": [0,1],
                                                                    "rh": [3,4]}},
                                                                    "Hippocampus": {"lh": [2],"rh": [5]}}}
-                raise ValueError(f"""
+                raise ValueError(textwrap.dedent(f"""
                                  {error_message}.
                                  These subkeys are needed for plotting. Please reassign `parcel_approach` using
-                                 `self.parcel_approach` amd refer to the example structure: {custom_example}""")
+                                 `self.parcel_approach` amd refer to the example structure: {custom_example}"""))
         if call  == "TimeseriesExtractor" and not os.path.isfile(parcel_approach["Custom"]["maps"]):
             raise ValueError("Please specify the location to the custom parcellation to be used.")